backend/requirements.txt
fastapi>=0.110
starlette>=0.36
uvicorn[standard]>=0.30
itsdangerous>=2.2
pydantic[email]==2.9.2
pandas==2.2.2
numpy==1.26.4
psycopg[binary]==3.2.3
sqlalchemy==2.0.34
alembic==1.13.2
redis==5.0.8
celery==5.4.0
qdrant-client==1.11.0
httpx==0.27.2
requests==2.32.3
python-dotenv==1.0.1
pytest==8.3.2
fakeredis==2.25.1
ruff==0.6.9
pydantic-settings==2.4.0
psycopg2-binary==2.9.9
openai>=1.42.0
authlib==1.3.2
passlib[bcrypt]>=1.7.4
tenacity>=8.2.0
aiohttp>=3.8.0
pyjwt==2.9.0

